dataset: Introduce the StanfordCars data module
Introduce the StanfordCars LightningDataModule. The Stanford Cars dataset
contains 16,185 images of 196 classes of cars. The data is split into 8,144
training images and 8,041 testing images, where each class has been split
roughly in a 50-50 split.

The dataset is wrapped as a LightningDataModule so the library can
leverage the features of PyTorch Lightning to train models and iterate
fast.

Signed-off-by: Dimitris Poulopoulos <dimpo@arrikto.com>