datasets: Introduce the FlowersData dataset
Oxford 102 Flower is an image classification dataset consisting of 102
flower categories. The flowers were chosen to be flowers commonly occurring
in the United Kingdom. Each class consists of between 40 and 258 images.

The dataset is wrapped as a LightningDataModule so the library can
leverage the features of PyTorch Lightning to train models and iterate
fast.

Signed-off-by: Dimitris Poulopoulos <dimpo@arrikto.com>